% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{This function is used to check and verify the input data to the package.
The package need a univariate time series as input.
This function keeps the first 2 columns, 1st is renamed as time and other is renamed as value.}
\usage{
check_input(df, dt_format, time, value)
}
\arguments{
\item{df}{Input data.}

\item{dt_format}{The format of the time column in the input data.}

\item{time}{Name of the column to use as time.}

\item{value}{Name of column to use as value(observations).}
}
\value{
Data containing 2 columns, time and value. Time column is POSIX object and value is double
}
\description{
This function is used to check and verify the input data to the package.
The package need a univariate time series as input.
This function keeps the first 2 columns, 1st is renamed as time and other is renamed as value.
}
