% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_timestamps.R
\name{find_dif}
\alias{find_dif}
\title{Helper function to find the time difference between two given timestamps.}
\usage{
find_dif(time1, time2)
}
\arguments{
\item{time1}{POSIXt or Date object.}

\item{time2}{POSIXt or Date object.}
}
\value{
String, specifying the time interval between \code{time1} and \code{time2}.
It contains a integer and the unit, for e.g., \emph{5 weeks}, \emph{6 months},
\emph{14 hours}, etc.
}
\description{
Helper function to find the time difference between two given timestamps.
}
