% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_report.R
\name{gen.report}
\alias{gen.report}
\title{Generate a report}
\usage{
gen.report(obj)
}
\arguments{
\item{obj}{A \emph{cleanTS} object.}
}
\description{
\code{gen.report()} generates a report of the entire process and the changes made
to the original data.
}
