% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{impute}
\alias{impute}
\title{This function handles missing values in the data.}
\usage{
impute(dt, methods)
}
\arguments{
\item{dt}{Input data}

\item{methods}{Imputation methods}
}
\description{
This function handles missing values in the data.
}
