% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_interval.R, R/missing_values.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{animate}
\alias{anim_save}
\alias{na_interpolation}
\alias{na_kalman}
\alias{na_locf}
\alias{na_ma}
\alias{na_mean}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gganimate}{\code{\link[gganimate]{anim_save}}, \code{\link[gganimate]{animate}}}

  \item{imputeTS}{\code{\link[imputeTS]{na_interpolation}}, \code{\link[imputeTS]{na_kalman}}, \code{\link[imputeTS]{na_locf}}, \code{\link[imputeTS]{na_ma}}, \code{\link[imputeTS]{na_mean}}}
}}

