% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_interval.R
\name{animate_interval}
\alias{animate_interval}
\title{Generate animated plot}
\usage{
animate_interval(obj, interval)
}
\arguments{
\item{obj}{A \emph{cleanTS} object.}

\item{interval}{A numeric or character, specifying the viewing interval.}
}
\value{
A list containing:
\itemize{
\item animation: A \code{gganim} object.
\item nstates: The number of states in the animation.
}
}
\description{
\code{animate_interval()} creates an animated plot using a \code{cleanTS} object
and a interval.
}
\examples{
\dontrun{
  anim <- animate_interval(cts, interval = "1 week")
}

}
