% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{detect_outliers}
\alias{detect_outliers}
\title{This function detects outliers in the data.}
\usage{
detect_outliers(dt, replace_outlier, imp_methods)
}
\arguments{
\item{dt}{Input data}

\item{replace_outlier}{Should the detected outliers be removed and imputated using the given imputation methods.}

\item{imp_methods}{Imputation methods}
}
\description{
This function detects outliers in the data.
}
