% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{impute}
\alias{impute}
\title{Handle missing values in the data}
\usage{
impute(dt, methods)
}
\arguments{
\item{dt}{A data.table.}

\item{methods}{The imputation methods to be used.}
}
\value{
A data.table with missing data imputed, and the imputation errors.
}
\description{
This function handles missing values in the data. It compares various
imputation methods and finds the best one for imputation.
}
