% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_plot.R
\name{interact_plot}
\alias{interact_plot}
\title{Create interactive plot}
\usage{
interact_plot(obj, interval)
}
\arguments{
\item{obj}{A \emph{cleanTS} object.}

\item{interval}{A numeric or character, specifying the viewing interval.}
}
\description{
Interactive plot is similar to the animated plot, but gives the used some
control over the animation. It runs a shinyApp instead of creating a GIF.
}
\examples{
\dontrun{
# Using the same data used in `cleanTS()` function example.
interact_plot(cts, interval = "1 week")
}

}
