% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergecsv.R
\name{mergecsv}
\alias{mergecsv}
\title{Merge Multiple CSV files}
\usage{
mergecsv(path, formats)
}
\arguments{
\item{path}{Path to the folder.}

\item{formats}{Datetime formats.}
}
\value{
Merged \code{data.table}.
}
\description{
\code{mergecsv()} takes a folder containing CSV files and merges them into a
single \emph{data.table}. It is assumed that the first column of all the CSVs
contains the timestamps.
}
