% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTS.R
\name{print.cleanTS}
\alias{print.cleanTS}
\title{Print a \emph{cleanTS} object}
\usage{
\method{print}{cleanTS}(x, ...)
}
\arguments{
\item{x}{cleanTS object}

\item{...}{Other arguments}
}
\description{
Print method for \code{cleanTS} class.
}
\examples{
\dontrun{
# Using the same data as in `cleanTS()` function example.
cts <- cleanTS(data, "my")
print(cts)
}

}
